+++ config/routes.rb
@@ -1,3 +1,5 @@
 RailsUrlShortener::Engine.routes.draw do
-  get '/:key', to: 'urls#show'
+  constraints RailsUrlShortener::HostConstraint do
+    get '/:key', to: 'urls#show'
+  end
 end